+++ src/backend/msal/authConfig.ts
@@ -1,18 +1,14 @@
-const { LogLevel } = require("@azure/msal-node");
-const path = require("path");
-const { cachePlugin } = require("./CachePlugin");
+import { LogLevel } from "@azure/msal-node";
 
 const AAD_ENDPOINT_HOST = "https://login.microsoftonline.com/"; // include the trailing slash
 
-const cachePath = path.resolve(__dirname, "./token.json");
-
-const msalConfig = {
+export const msalConfig = {
   auth: {
     clientId: "f1ffa820-44a2-43da-9016-d3302c89c36a",
     authority: `${AAD_ENDPOINT_HOST}/consumers`,
   },
   cache: {
-    cachePlugin: cachePlugin(cachePath),
+    cachePlugin: null,
   },
   system: {
     loggerOptions: {
@@ -27,7 +23,7 @@ const msalConfig = {
 
 const GRAPH_ENDPOINT_HOST = "https://graph.microsoft.com/"; // include the trailing slash
 
-const protectedResources = {
+export const protectedResources = {
   graphMe: {
     endpoint: `${GRAPH_ENDPOINT_HOST}v1.0/me`,
     scopes: ["User.Read", "Files.Read", "offline_access"],
@@ -50,13 +46,6 @@ const protectedResources = {
   },
 };
 
-const getRelativeResourceUrl = (url) => {
+export const getRelativeResourceUrl = (url: string) => {
   return url.replace(`${GRAPH_ENDPOINT_HOST}v1.0`, "");
-};
-
-module.exports = {
-  msalConfig: msalConfig,
-  cachePath: cachePath,
-  protectedResources: protectedResources,
-  getRelativeResourceUrl: getRelativeResourceUrl,
 };